requirements.txt
cooler==0.8.11
tqdm==4.64.0
numpy==1.21.5
scipy==1.9.1
pandas==0.25.3
matplotlib==3.1.3
joblib==1.1.0
scikit-learn==1.1.1
torchmetrics
kcounter==0.1.1
pyfastx==0.8.4